Move initialization functions into library
